backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.3.2
pymongo>=4.6.0
python-dotenv>=1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.10.0
pydantic[email]>=2.10.0
email-validator>=2.0.0
